# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Gapak API
  version: 1.0.0
  description: |
    Gapak is a privacy-first social network built around trust, layered identity,
    controlled visibility, secure media, live experiences, and time-based social interaction.

    Privacy is a core mechanic rather than a decorative setting. The API is designed for
    trust-aware access rules, metadata minimization, secure media delivery, and live social primitives.
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    accessCookie:
      type: apiKey
      in: cookie
      name: gapak_at
  schemas:
...
    MessageRequest:
      type: object
      required: [clientMessageId, envelopeType, ciphertext, nonce, senderKeyId]
      properties:
        clientMessageId: { type: string }
        envelopeType: { type: string, enum: [TEXT, ATTACHMENT, KEY_EXCHANGE, SYSTEM] }
        ciphertext: { type: string }
        nonce: { type: string }
        senderKeyId: { type: string }
        attachmentManifest: { type: object, additionalProperties: true }
        metadata: { type: object, additionalProperties: true }
...
paths:
...
  /api/v1/chats:
    get:
      tags: [Chats]
      security: [{ bearerAuth: [] }]
      summary: List chats
      responses:
        '200': { description: Chat list }
...
  /api/v1/chats/pre-key-bundles/{userId}:
    get:
      tags: [Chats]
      security: [{ bearerAuth: [] }]
      summary: Get the authoritative E2EE pre-key bundle for a recipient user
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Pre-key bundle returned in the standard success envelope }
        '400': { description: Invalid userId }
        '401': { description: Authentication required }
        '403': { description: Recipient is not visible to the current user }
        '404': { description: Recipient has no usable trusted device bundle }
  /api/v1/chats/{chatId}/messages:
    get:
      tags: [Chats]
      security: [{ bearerAuth: [] }]
      summary: List encrypted message envelopes
      parameters:
        - in: path
          name: chatId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Message list }
    post:
      tags: [Chats]
      security: [{ bearerAuth: [] }]
      summary: Store encrypted message envelope without plaintext dependency
      parameters:
        - in: path
          name: chatId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MessageRequest' }
      responses:
        '201': { description: Message stored }
...